http: move http test utilities into http library
The previous location, test_utils library, results in circular
dependencies. test_utils depends on essentially nothing except seastar,
but having http bits in test_utils library means that it then needs to
depend on our http library, for instance. however, most libraries in
turn depend no test_utils.
