+++ src/v/archival/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ else()
       BINARY_NAME test_archival_service
       SOURCES service_fixture.cc ntp_archiver_test.cc segment_reupload_test.cc ntp_archiver_reupload_test.cc retention_strategy_test.cc
       DEFINITIONS BOOST_TEST_DYN_LINK
-      LIBRARIES v::seastar_testing_main v::application Boost::unit_test_framework v::archival v::storage_test_utils v::cloud_roles
+      LIBRARIES v::seastar_testing_main v::application Boost::unit_test_framework v::archival v::storage_test_utils v::cloud_roles v::http_test_utils
       ARGS "-- -c 1"
       LABELS archival
     )
