+++ src/v/cloud_roles/tests/fetch_credentials_tests.cc
@@ -11,8 +11,8 @@
 #include "bytes/iobuf.h"
 #include "cloud_roles/refresh_credentials.h"
 #include "cloud_roles/tests/test_definitions.h"
+#include "http/tests/http_imposter.h"
 #include "test_utils/fixture.h"
-#include "test_utils/http_imposter.h"
 
 #include <seastar/core/file.hh>
