+++ src/v/cluster/tests/CMakeLists.txt
@@ -53,7 +53,7 @@ rp_test(
   UNIT_TEST
   BINARY_NAME test_${test_name}
   SOURCES ${cluster_test_src}
-  LIBRARIES v::seastar_testing_main v::application v::storage_test_utils v::cluster
+  LIBRARIES v::seastar_testing_main v::application v::storage_test_utils v::cluster v::http_test_utils
   LABELS cluster
 )
 endforeach()
