+++ src/v/http/tests/http_imposter.h
@@ -10,9 +10,9 @@
 
 #pragma once
 
+#include "http/tests/registered_urls.h"
 #include "net/unresolved_address.h"
 #include "seastarx.h"
-#include "test_utils/registered_urls.h"
 #include "vassert.h"
 
 #include <seastar/core/sstring.hh>
