+++ src/v/http/tests/http_imposter_test.cc
@@ -10,8 +10,8 @@
 
 #include "bytes/iobuf_parser.h"
 #include "http/client.h"
+#include "http/tests/http_imposter.h"
 #include "test_utils/fixture.h"
-#include "test_utils/http_imposter.h"
 
 #include <boost/test/unit_test.hpp>
