+++ src/v/http/tests/registered_urls.h
