+++ src/v/test_utils/CMakeLists.txt
@@ -1,6 +1,6 @@
 v_cc_library(
   NAME seastar_testing_main
-  SRCS seastar_testing_main.cc http_imposter.cc registered_urls.cc
+  SRCS seastar_testing_main.cc
   DEPS Seastar::seastar_testing)
 
 add_subdirectory(tests)
